requirements.txt
aiolimiter==1.2.1
anyio==4.8.0
APScheduler==3.11.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
MouseInfo==0.1.3
pillow==11.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-telegram-bot==21.10
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.2
pywhatkit==5.4
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
tornado==6.4.2
typing_extensions==4.12.2
tzlocal==5.2
urllib3==1.26.15
Werkzeug==3.1.3
wikipedia==1.4.0
